Rescue the reencoding of messages but omit non-desirable parts.
Some bounces use 'message/report' content-type for reporting the error.
And, then, embed the original message as 'text/*'.  Odoo prefers the
'text/' over the bounce message and what we get as the body of the
message has lost the important data: the bounce reason.

Saving this information for later can help a bit.

However, we should actually modify Odoo's behavior to include
'message/*' parts in the body.