Update the router/transport integrators to the new API.
The router and transport will be updated soon to eliminate the cr, uid, and
context arguments.  Though we will provide deprecation warnings.

This commit solely updates the methods that integrate the routers/transports
into Odoo's mail receiving/sending mechanism.

The idea is NOT TO BREAK Odoo 8 and our current routers.  But allowed them to
be updated.

OpenERP 7 is no longer supported.