Stop using the currently unusable xoeuf's Registry.
There are changes in Odoo 10's Registry that are not yet ported to xoeuf's.
This makes it harder to work with the xoeuf's Registry in a portable way.

So it's best to keep things simple by using the new API (we no longer support
OpenERP 7) and pass whole recordset to the is_router_installed function.