Redesign test suite.
When mocking `query` we should control the return value at the test, so that
we can assert if other methods should be called.

I had to split the tests in order to use `patch.multiple` as a decorator at
the class level.  Using it in a single class as decorator for its methods
shadows the `self` and fails with:

   TypeError: test_...() takes exactly 3 arguments (2 given)