Fix test suite.
I could not make `patch.multiple` work with the hand-crafted mocks.

The transport tests fails in Odoo 10.  While debugging we uncover a bug in
Odoo 9+'s tests [1]_.

Now, there seems to be another one (in the Odoo 10 tests): When our tests run,
the `send_email` method of model 'ir.mail_server' is not our implementation.
When using the shell, our method is present and not by-passed.