Fix transports test to deal with patched send_email in addon 'mail'.
The pair of methods _patch_method and _revert_method are ill-defined.  In this
case: the tests of the addon 'mail' patch the method `send_email` of
ir.mail_server, and revert, this causes the Model Class to be modified by
setting the 'original' code of send_email *in the Model Class*.  Since the
original method does not calls a super our send_email delivering via
transports is bypassed.

The Model Class should be empty: so we simply remove the method.